Bump github-release from 0.7.2 to 0.8.1
And from its old ownership (aktau) to the Github Organization that
it's owned by now (github-release).

It seems the new versions are packaged differently: it used to be a
tar.bz2 file with the executable in a `bin/` directory. The new
release is just a straight up `bz2` of the executable, so the way it
is unpacked has changed slightly.