Allow catapult-publish-lambda to handle multiple regions at once.
This is backwards compatible based on the run-type of the
application. Any lambda using run-type `lambda/<region>` will do the
same thing as before.

However, if the run-type is just `lambda`, with no region, then the
set of regions will be pulled from either the AWS_REGIONS env var set
by CI, or else default to the four main US regions.